+++ src/common/settings.ts
@@ -476,6 +476,8 @@ export class Settings {
 
     public allowCommentsFromCommand: boolean = false;
     public saveTriggerStateChanges: boolean = true;
+    public groupProfileSaves: boolean = false;
+    public groupProfileSaveDelay: number = 20000;
 
     public windows = {};
     public buttons = {