+++ changelog.md
@@ -17,6 +17,7 @@
     - Add [Block Inline Comment](docs/preferences.md#scripting--special-characters) disable block comment support    
     - Add [Save Trigger State Changes](docs/preferences.md#advanced--profiles) Add option to control if state changes are saved every time they change
     - New Advanced sub page for all profile related preferences.
+    - Add [Group profile saves](docs/preferences.md#advanced--profiles) and [Group profile save delay](docs/preferences.md#advanced--profiles) to allow grouping of profile saves
   - [Triggers](docs/profiles.md#triggers)
     - Add `Loop Expression` trigger type
     - Add `ReParse` and `ReParse Pattern` sub type triggers